+++ GRDB/Core/Configuration.swift
@@ -20,6 +20,53 @@ public struct Configuration {
     /// Default: false
     public var readonly: Bool = false
     
+    /// The database label.
+    ///
+    /// You can query this label at runtime:
+    ///
+    ///     var configuration = Configuration()
+    ///     configuration.label = "MyDatabase"
+    ///     let dbQueue = try DatabaseQueue(path: ..., configuration: configuration)
+    ///
+    ///     try dbQueue.read { db in
+    ///         print(db.configuration.label) // Prints "MyDatabase"
+    ///     }
+    ///
+    /// The database label is also used to name the various dispatch queues
+    /// created by GRDB, visible in debugging sessions and crash logs. However
+    /// those dispatch queue labels are intended for debugging only. Their
+    /// format may change between GRDB releases. Applications should not depend
+    /// on the GRDB dispatch queue labels.
+    ///
+    /// If the database label is nil, the current GRDB implementation uses the
+    /// following dispatch queue labels:
+    ///
+    /// - `GRDB.DatabaseQueue`: the (unique) dispatch queue of a DatabaseQueue
+    /// - `GRDB.DatabasePool.writer`: the (unique) writer dispatch queue of
+    ///   a DatabasePool
+    /// - `GRDB.DatabasePool.reader.N`, where N is 1, 2, ...: one of the reader
+    ///   dispatch queue(s) of a DatabasePool. N grows with the number of SQLite
+    ///   connections: it may get bigger than the maximum number of concurrent
+    ///   readers, as SQLite connections get closed and new ones are opened.
+    /// - `GRDB.DatabasePool.snapshot.N`: the dispatch queue of a
+    ///   DatabaseSnapshot. N grows with the number of snapshots.
+    ///
+    /// If the database label is not nil, for example "MyDatabase", the current
+    /// GRDB implementation uses the following dispatch queue labels:
+    ///
+    /// - `MyDatabase`: the (unique) dispatch queue of a DatabaseQueue
+    /// - `MyDatabase.writer`: the (unique) writer dispatch queue of
+    ///   a DatabasePool
+    /// - `MyDatabase.reader.N`, where N is 1, 2, ...: one of the reader
+    ///   dispatch queue(s) of a DatabasePool. N grows with the number of SQLite
+    ///   connections: it may get bigger than the maximum number of concurrent
+    ///   readers, as SQLite connections get closed and new ones are opened.
+    /// - `MyDatabase.snapshot.N`: the dispatch queue of a
+    ///   DatabaseSnapshot. N grows with the number of snapshots.
+    ///
+    /// The default label is nil.
+    public var label: String? = nil
+    
     /// A function that is called on every statement executed by the database.
     ///
     /// Default: nil