+++ Tests/GRDBTests/DatabaseSnapshotTests.swift
@@ -119,4 +119,53 @@ class DatabaseSnapshotTests: GRDBTestCase {
             XCTAssertEqual(try String.fetchAll(db, "SELECT text FROM items ORDER BY text COLLATE reverse"), ["c", "b", "a"])
         }
     }
+    
+    func testDefaultLabel() throws {
+        let dbPool = try makeDatabasePool()
+        
+        let snapshot1 = try dbPool.makeSnapshot()
+        snapshot1.unsafeRead { db in
+            XCTAssertEqual(db.configuration.label, nil)
+            
+            // This test CAN break in future releases: the dispatch queue labels
+            // are documented to be a debug-only tool.
+            let label = String(utf8String: __dispatch_queue_get_label(nil))
+            XCTAssertEqual(label, "GRDB.DatabasePool.snapshot.1")
+        }
+        
+        let snapshot2 = try dbPool.makeSnapshot()
+        snapshot2.unsafeRead { db in
+            XCTAssertEqual(db.configuration.label, nil)
+            
+            // This test CAN break in future releases: the dispatch queue labels
+            // are documented to be a debug-only tool.
+            let label = String(utf8String: __dispatch_queue_get_label(nil))
+            XCTAssertEqual(label, "GRDB.DatabasePool.snapshot.2")
+        }
+    }
+    
+    func testCustomLabel() throws {
+        dbConfiguration.label = "Toreador"
+        let dbPool = try makeDatabasePool()
+        
+        let snapshot1 = try dbPool.makeSnapshot()
+        snapshot1.unsafeRead { db in
+            XCTAssertEqual(db.configuration.label, "Toreador")
+            
+            // This test CAN break in future releases: the dispatch queue labels
+            // are documented to be a debug-only tool.
+            let label = String(utf8String: __dispatch_queue_get_label(nil))
+            XCTAssertEqual(label, "Toreador.snapshot.1")
+        }
+        
+        let snapshot2 = try dbPool.makeSnapshot()
+        snapshot2.unsafeRead { db in
+            XCTAssertEqual(db.configuration.label, "Toreador")
+            
+            // This test CAN break in future releases: the dispatch queue labels
+            // are documented to be a debug-only tool.
+            let label = String(utf8String: __dispatch_queue_get_label(nil))
+            XCTAssertEqual(label, "Toreador.snapshot.2")
+        }
+    }
 }